fix the reviewer's comments
