+++ src/services/actions/ws-actions-auth.js
@@ -5,3 +5,12 @@ export const WS_CONNECTION_CLOSED_AUTH = 'WS_CONNECTION_CLOSED_AUTH';
 export const WS_GET_MESSAGE_AUTH = 'WS_GET_MESSAGE_AUTH';
 export const WS_SEND_MESSAGE_AUTH = 'WS_SEND_MESSAGE_AUTH';
 export const WS_USER_NAME_UPDATE_AUTH = 'WS_USER_NAME_UPDATE_AUTH';
+
+export const wsActionsAuth = {
+	wsInit: WS_CONNECTION_START_AUTH,
+	wsSendMessage: WS_SEND_MESSAGE_AUTH,
+	onOpen: WS_CONNECTION_SUCCESS_AUTH,
+	onClose: WS_CONNECTION_CLOSED_AUTH,
+	onError: WS_CONNECTION_ERROR_AUTH,
+	onMessage: WS_GET_MESSAGE_AUTH
+};
