+++ src/services/actions/ws-actions.js
@@ -6,10 +6,12 @@ export const WS_GET_MESSAGE = 'WS_GET_MESSAGE';
 export const WS_SEND_MESSAGE = 'WS_SEND_MESSAGE';
 export const WS_USER_NAME_UPDATE = 'WS_USER_NAME_UPDATE';
 
-export const WS_CONNECTION_START_AUTH = 'WS_CONNECTION_START_AUTH';
-export const WS_CONNECTION_SUCCESS_AUTH = 'WS_CONNECTION_SUCCESS_AUTH';
-export const WS_CONNECTION_ERROR_AUTH = 'WS_CONNECTION_ERROR_AUTH';
-export const WS_CONNECTION_CLOSED_AUTH = 'WS_CONNECTION_CLOSED_AUTH';
-export const WS_GET_MESSAGE_AUTH = 'WS_GET_MESSAGE_AUTH';
-export const WS_SEND_MESSAGE_AUTH = 'WS_SEND_MESSAGE_AUTH';
-export const WS_USER_NAME_UPDATE_AUTH = 'WS_USER_NAME_UPDATE_AUTH';
+
+export const wsActions = {
+	wsInit: WS_CONNECTION_START,
+	wsSendMessage: WS_SEND_MESSAGE,
+	onOpen: WS_CONNECTION_SUCCESS,
+	onClose: WS_CONNECTION_CLOSED,
+	onError: WS_CONNECTION_ERROR,
+	onMessage: WS_GET_MESSAGE
+};
